requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
httpx>=0.27.0,<0.28.0
pydantic>=2.7.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
boto3>=1.40.45
python-multipart>=0.0.20
pydantic-settings>=2.11.0
google-generativeai>=0.3.0
mangum>=0.17.0
